# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Resilience API",
    "description": "Deterministic application boundary for the Resilience prototype.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/scheme-navigator/questionnaire": {
      "get": {
        "tags": [
          "scheme-navigator"
        ],
        "summary": "Get Questionnaire",
        "description": "Returns only the questions actually needed by the currently loaded\nscheme rules, in a stable display order.",
        "operationId": "get_questionnaire_api_v1_scheme_navigator_questionnaire_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuestionnaireField"
                  },
                  "type": "array",
                  "title": "Response Get Questionnaire Api V1 Scheme Navigator Questionnaire Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/scheme-navigator/evaluate": {
      "post": {
        "tags": [
          "scheme-navigator"
        ],
        "summary": "Evaluate",
        "description": "Evaluates the submitted answers against every loaded scheme rule.\n\nAny field a rule needs that is absent from ``answers`` yields a\n``missing_information`` result for that scheme rather than a guess.",
        "operationId": "evaluate_api_v1_scheme_navigator_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/scheme-navigator/explain": {
      "post": {
        "tags": [
          "scheme-navigator"
        ],
        "summary": "Explain Result",
        "description": "Rephrases an already-decided result in plain language.\n\nThe status is decided by the evaluator and only described here. If the\nmodel is unconfigured or unreachable, this returns a deterministic\nexplanation rather than an error.",
        "operationId": "explain_result_api_v1_scheme_navigator_explain_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplanationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExplanationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/scheme-navigator/chat": {
      "post": {
        "tags": [
          "scheme-navigator"
        ],
        "summary": "Chat Turn",
        "description": "Answers a scoped question about the loaded schemes and this app.\n\nDegrades to a canned reply when the model is unconfigured or\nunreachable, so the chat panel never surfaces a server error.",
        "operationId": "chat_turn_api_v1_scheme_navigator_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatMessage": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 1500,
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ChatMessage"
      },
      "ChatRequest": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Messages"
          },
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 32,
            "title": "Answers"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SchemeResult"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Results"
          }
        },
        "type": "object",
        "title": "ChatRequest",
        "description": "A chatbot turn.\n\nUnlike the explainer, this deliberately carries the user's ``answers``:\nthe assistant is expected to field questions like \"why didn't I match\nWIS?\", which cannot be answered without them. That is a conscious\nprivacy trade -- see documentation/features/scheme-navigator.md."
      },
      "ChatResponse": {
        "properties": {
          "reply": {
            "type": "string",
            "title": "Reply"
          },
          "is_ai_generated": {
            "type": "boolean",
            "title": "Is Ai Generated"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "reply",
          "is_ai_generated",
          "generated_at"
        ],
        "title": "ChatResponse"
      },
...
      "EvaluationRequest": {
        "properties": {
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 32,
            "title": "Answers"
          }
        },
        "type": "object",
        "title": "EvaluationRequest"
      },
      "EvaluationResponse": {
        "properties": {
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SchemeResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "generated_at",
          "results"
        ],
        "title": "EvaluationResponse"
      },
      "ExplanationRequest": {
        "properties": {
          "rule_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Rule Id"
          },
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "maxProperties": 32,
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "rule_id"
        ],
        "title": "ExplanationRequest",
        "description": "Asks for a plain-language explanation of an already-decided result.\n\nThe caller sends back a ``SchemeResult`` the evaluator produced. The\nexplainer never re-derives the status, so the request carries the\noutcome rather than the answers that led to it."
      },
      "ExplanationResponse": {
        "properties": {
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "next_steps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next Steps"
          },
          "source_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Source Urls"
          },
          "is_ai_generated": {
            "type": "boolean",
            "title": "Is Ai Generated"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          }
        },
        "type": "object",
        "required": [
          "summary",
          "is_ai_generated",
          "generated_at"
        ],
        "title": "ExplanationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionnaireField": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "field_type": {
            "type": "string",
            "enum": [
              "number",
              "boolean",
              "select",
              "date"
            ],
            "title": "Field Type"
          },
          "help_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Help Text"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "options": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SelectOption"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Options"
          },
          "min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Value"
          },
          "max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Value"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "field_type"
        ],
        "title": "QuestionnaireField",
        "description": "One prompt shown to the user. Derived from whichever scheme rules\nactually reference it, not hand-written per scheme."
      },
...
      "SchemeResult": {
        "properties": {
          "rule_id": {
            "type": "string",
            "title": "Rule Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "agency": {
            "type": "string",
            "title": "Agency"
          },
          "status": {
            "$ref": "#/components/schemas/SchemeStatus"
          },
          "matched_facts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Facts"
          },
          "unmatched_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Unmatched Reasons"
          },
          "missing_fields": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Fields"
          },
          "official_source_url": {
            "type": "string",
            "title": "Official Source Url"
          },
          "application_url": {
            "type": "string",
            "title": "Application Url"
          },
          "last_reviewed_date": {
            "type": "string",
            "format": "date",
            "title": "Last Reviewed Date"
          },
          "simplified_note": {
            "type": "string",
            "title": "Simplified Note"
          }
        },
        "type": "object",
        "required": [
          "rule_id",
          "name",
          "agency",
          "status",
          "official_source_url",
          "application_url",
          "last_reviewed_date",
          "simplified_note"
        ],
        "title": "SchemeResult"
      },
      "SchemeStatus": {
        "type": "string",
        "enum": [
          "matched",
          "not_matched",
          "missing_information"
        ],
        "title": "SchemeStatus"
      },
      "SelectOption": {
        "properties": {
          "value": {
            "type": "string",
            "title": "Value"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "value",
          "label"
        ],
        "title": "SelectOption"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}